+++ packages/mockopampserver/lib/mockopampserver.js
@@ -184,7 +184,7 @@ class MockOpAMPServer {
         this._port = opts.port ?? DEFAULT_PORT;
         this._endpointPath = DEFAULT_ENDPOINT_PATH;
         if (opts.agentConfigMap) {
-            this._setAgentConfigMap(opts.agentConfigMap);
+            this.setAgentConfigMap(opts.agentConfigMap);
         }
         this._server = http.createServer(this._onRequest.bind(this));
         this._started = false;
@@ -210,7 +210,32 @@ class MockOpAMPServer {
         }
     }
 
-    _setAgentConfigMap(agentConfigMap) {
+    /**
+     * Set the data used by the server to provide `remoteConfig` to agents.
+     *
+     * `agentConfigMap` is of the form:
+     *      {
+     *          configMap: {
+     *              // Zero or more entries in `configMap`.
+     *              'some-key': {
+     *                  body: <Uint8Array of config file content>,
+     *                  contentType: <string>
+     *              }
+     *          }
+     *      }
+     *
+     * Example usage:
+     *      const config = { deactivate_all_instrumentations: 'true' };
+     *      opampServer.setAgentConfigMap({
+     *        configMap: {
+     *          elastic: {
+     *            body: Buffer.from(JSON.stringify(config), 'utf8'),
+     *            contentType: 'application/json',
+     *          }
+     *        }
+     *      });
+     */
+    setAgentConfigMap(agentConfigMap) {
         this._agentConfigMap = agentConfigMap;
         this._agentConfigMapHash = hashAgentConfigMap(agentConfigMap);
     }
@@ -362,7 +387,7 @@ class MockOpAMPServer {
         let agentConfigMap = {configMap: {}};
         const finish = () => {
             log.trace({agentConfigMap}, 'SetAgentConfigMap');
-            this._setAgentConfigMap(agentConfigMap);
+            this.setAgentConfigMap(agentConfigMap);
             res.writeHead(204);
             res.end();
             log.debug({req, res}, 'test API request: SetAgentConfigMap');